# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de gestión de alumnos",
    "description": "Documento generado automaticamente por swagger-autogen",
    "version": "1.0.0"
  },
  "host": "lzrlgb5t-3000.brs.devtunnels.ms",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "BypassTunnelAlert": {
      "type": "apiKey",
      "in": "header",
      "name": "X-Tunnel-Skip-AntiPhishing-Threshold",
      "description": "Escribe \"true\" para saltar el bloqueo de Dev Tunnels"
    }
  },
  "paths": {
...
    "/socios/": {
      "get": {
        "tags": [
          "Socios"
        ],
        "description": "Endpoint para listar los socios y consultar por suscripto (true/false)",
        "parameters": [
          {
            "name": "suscripto",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Socios"
        ],
        "description": "Endpoint para registrar un Nuevo Socio. Todos los campos son String, excepto <suscripto> que es boolean (true/false)",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "example": "any"
                },
                "dni": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "suscripto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/socios/{id}": {
      "get": {
        "tags": [
          "Socios"
        ],
        "description": "Endpoint para buscar a un Socio con su ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Socios"
        ],
        "description": "Endpoint para Actualizar solo el contacto (email) y/o membresia (suscripto), de un socio por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "suscripto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Socios"
        ],
        "description": "Endpoint para Eliminar a un socio por ID ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/sanciones/": {
      "get": {
        "tags": [
          "Sanciones"
        ],
        "description": "Endpoint para listar todas las sanciones ",
        "parameters": [
          {
            "name": "pagada",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Sanciones"
        ],
        "description": "Endpoint para registrar una nueva Sanción ",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "socioId": {
                  "example": "any"
                },
                "motivo": {
                  "example": "any"
                },
                "monto": {
                  "example": "any"
                },
                "pagada": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/sanciones/{id}": {
      "get": {
        "tags": [
          "Sanciones"
        ],
        "description": "Endpoint para mostrar una sancion por ID ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Sanciones"
        ],
        "description": "Endpoint para eliminar una Sanción por ID ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "security": [
    {
      "BypassTunnelAlert": []
    }
  ]
}